requirements.txt
asgiref==3.7.2
black==22.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
delocate==0.10.7
Django==5.0.2
django-cors-headers==4.3.1
django-devserver==0.8.0
django-filter==24.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
flake8==4.0.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
mccabe==0.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
olefile==0.47
packaging==23.2
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyHamcrest==2.1.0
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.0
